+++ README.md
@@ -14,6 +14,7 @@
 * 지정과제
   - 기업형 : TmaxOS / 클라우드바리스타 / 매니코어SW
   - 사회문제형 : 국민생활과 밀접한 사회문제해결을 위한 공개SW 프로젝트 개발. 환경/사회적 약자/감염병과 관련된 사회문제를 과제로 선정 후 해결방안 제시
+
 ## 일정
 |구분|기간|상세내용|
 |:---:|:---:|:---:|
@@ -25,6 +26,31 @@
 |멘토링|9.27(월) ~ 10.22(금)|1차 평가 합격자 대상 팀별 맞춤형 멘토링 제공|
 |2차 평가|10.25(월) ~ 11.5(금)|- 기능테스트: 실제 구현을 통한 출품작 검증 및 보완 <br/>- 라이선스 검증 : 출품작 라이선스 위반 사항 및 추돌 여부 검증 <br/>(11.9(화) 발표 평가)|
 | 시상식|11월 말(SW주간)|공개SW 주간 : 수상자 시상 및 발표 등|
+
+## 코드
+### [파이썬 인증 모듈](./servers/python)
+
+### [백엔드 인증 모듈]
+
+### 실행 방법
+1. `git clone https://github.com/AI-Jiny/OSS.git` 또는 다운로드 한다.
+2. [노드 사이트](https://nodejs.org/ko/download/)에서 노드 다운로드 후 설치
+3. cmd 창에서 `cd 파일위치`
+4. cmd 창에서 `npm i`
+5. cmd 창에서 `cd 파일위치/servers/python`
+6. cmd 창에서 `pip install -r requirements.txt`
+7. cmd 창에서 테스트용 프론트 엔드 실행
+```bash
+npm start
+```
+8. cmd 창에서 백엔드 실행
+```bash
+npm run server
+```
+9. 프론트 엔드에서 이름과 문서번호를 입력하여 테스트
+
+### 주의 사항
+  - 크롬 기반으로 실행 되기에 크롬 드라이버와 크롬 버전이 동일해야 함
 
 ## 추진기관
 - 주최 : 과학기술정보통신부
