+++ servers/python/requirments.txt
@@ -1,0 +1,7 @@
+certifi==2021.5.30
+urllib3==1.26.5
+numpy==1.20.3
+opencv-python==4.5.2.54
+robotframework==4.0.3
+selenium==3.141.0
+wincertstore==0.2
